+++ module/Activity/config/module.config.php
@@ -71,13 +71,14 @@ return [
                 ],
                 'priority' => 100
             ],
-            'organizer_activity' => [
+            'activity_admin' => [
+                'priority' => 100,
                 'type' => 'Literal',
                 'options' => [
-                    'route' => '/activity/organizer',
+                    'route' => '/admin/activity',
                     'defaults' => [
                         '__NAMESPACE__' => 'Activity\Controller',
-                        'controller' => 'organizer',
+                        'controller' => 'admin',
                         'action' => 'view'
                     ],
                 ],
@@ -91,7 +92,7 @@ return [
                                 'page' => '[0-9]+',
                             ],
                             'defaults' => [
-                                'controller' => 'organizer',
+                                'controller' => 'admin',
                                 'action' => 'view'
                             ]
                         ]
@@ -101,7 +102,7 @@ return [
                         'options' => [
                             'route' => '/:id/email',
                             'defaults' => [
-                                'controller' => 'organizer',
+                                'controller' => 'admin',
                                 'action' => 'email',
                             ]
                         ]
@@ -111,7 +112,7 @@ return [
                         'options' => [
                             'route' => '/:id/export',
                             'defaults' => [
-                                'controller' => 'organizer',
+                                'controller' => 'admin',
                                 'action' => 'export',
                             ]
                         ]
@@ -121,7 +122,7 @@ return [
                         'options' => [
                             'route' => '/:id/export/pdf',
                             'defaults' => [
-                                'controller' => 'organizer',
+                                'controller' => 'admin',
                                 'action' => 'exportpdf',
                             ]
                         ]
@@ -131,7 +132,7 @@ return [
                         'options' => [
                             'route' => '/:id/update',
                             'defaults' => [
-                                'controller' => 'organizer',
+                                'controller' => 'admin',
                                 'action' => 'update'
                             ]
                         ]
@@ -163,13 +164,14 @@ return [
                 ]
 
             ],
-            'admin_activity' => [
+            'activity_admin_approval' => [
+                'priority' => 150,
                 'type' => 'Literal',
                 'options' => [
-                    'route' => '/admin/activity',
+                    'route' => '/admin/activity/approval',
                     'defaults' => [
                         '__NAMESPACE__' => 'Activity\Controller',
-                        'controller' => 'admin',
+                        'controller' => 'adminApproval',
                         'action' => 'queue'
                     ]
                 ],
@@ -180,7 +182,7 @@ return [
                         'options' => [
                             'route' => '/queue',
                             'defaults' => [
-                                'controller' => 'admin',
+                                'controller' => 'adminApproval',
                                 'action' => 'queue'
                             ]
                         ]
@@ -190,7 +192,7 @@ return [
                         'options' => [
                             'route' => '/queue/unapproved[/:page]',
                             'defaults' => [
-                                'controller' => 'admin',
+                                'controller' => 'adminApproval',
                                 'action' => 'queueUnapproved'
                             ]
                         ]
@@ -200,7 +202,7 @@ return [
                         'options' => [
                             'route' => '/queue/approved[/:page]',
                             'defaults' => [
-                                'controller' => 'admin',
+                                'controller' => 'adminApproval',
                                 'action' => 'queueApproved'
                             ]
                         ]
@@ -210,7 +212,7 @@ return [
                         'options' => [
                             'route' => '/queue/disapproved[/:page]',
                             'defaults' => [
-                                'controller' => 'admin',
+                                'controller' => 'adminApproval',
                                 'action' => 'queueDisapproved'
                             ]
                         ]
@@ -220,7 +222,7 @@ return [
                         'options' => [
                             'route' => '/view/[:id]',
                             'defaults' => [
-                                'controller' => 'admin',
+                                'controller' => 'adminApproval',
                                 'action' => 'view'
                             ]
                         ]
@@ -230,7 +232,7 @@ return [
                         'options' => [
                             'route' => '/proposal/[:id]',
                             'defaults' => [
-                                'controller' => 'admin',
+                                'controller' => 'adminApproval',
                                 'action' => 'viewProposal'
                             ]
                         ]
@@ -240,7 +242,7 @@ return [
                         'options' => [
                             'route' => '/proposal/[:id]/apply',
                             'defaults' => [
-                                'controller' => 'admin',
+                                'controller' => 'adminApproval',
                                 'action' => 'applyProposal'
                             ]
                         ]
@@ -250,7 +252,7 @@ return [
                         'options' => [
                             'route' => '/proposal/[:id]/revoke',
                             'defaults' => [
-                                'controller' => 'admin',
+                                'controller' => 'adminApproval',
                                 'action' => 'revokeProposal'
                             ]
                         ]
@@ -260,7 +262,7 @@ return [
                         'options' => [
                             'route' => '/approve/[:id]',
                             'defaults' => [
-                                'controller' => 'admin',
+                                'controller' => 'adminApproval',
                                 'action' => 'approve'
                             ]
                         ]
@@ -270,7 +272,7 @@ return [
                         'options' => [
                             'route' => '/disapprove/[:id]',
                             'defaults' => [
-                                'controller' => 'admin',
+                                'controller' => 'adminApproval',
                                 'action' => 'disapprove'
                             ]
                         ]
@@ -280,7 +282,7 @@ return [
                         'options' => [
                             'route' => '/reset/[:id]',
                             'defaults' => [
-                                'controller' => 'admin',
+                                'controller' => 'adminApproval',
                                 'action' => 'reset'
                             ]
                         ]
@@ -361,9 +363,9 @@ return [
     'controllers' => [
         'invokables' => [
             'Activity\Controller\Activity' => 'Activity\Controller\ActivityController',
-            'Activity\Controller\Admin' => 'Activity\Controller\AdminController',
+            'Activity\Controller\AdminApproval' => 'Activity\Controller\AdminApprovalController',
             'Activity\Controller\Api' => 'Activity\Controller\ApiController',
-            'Activity\Controller\Organizer' => 'Activity\Controller\OrganizerController',
+            'Activity\Controller\Admin' => 'Activity\Controller\AdminController',
             'Activity\Controller\ActivityCalendar' => 'Activity\Controller\ActivityCalendarController',
         ],
         'factories' => [
