+++ module/Activity/view/activity/activity/list.phtml
@@ -20,12 +20,12 @@
                 <div class="col-md-12">
                     <h4>
                         <?php if (!isset($proposals)): ?>
-                        <a href="<?php echo $this->url((isset($admin) ? 'admin_activity' : 'activity'). '/view', ['id' => $activity->getId()])?>">
+                        <a href="<?php echo $this->url((isset($admin) ? 'activity_admin_approval' : 'activity'). '/view', ['id' => $activity->getId()])?>">
                             <?= $this->escapeHtml($activity->getName() === null ? $activity->getNameEn() : $activity->getName()) ?>
                         </a>
                         <?php else: ?>
                         <?php $proposal = $proposals[$activity->getId()] ?>
-                        <a href="<?php echo $this->url('admin_activity/proposal', ['id' => $proposal->getId()]) ?>">
+                        <a href="<?php echo $this->url('activity_admin_approval/proposal', ['id' => $proposal->getId()]) ?>">
                             <?= $this->translate('Proposal to update into: ') . $this->escapeHtml($activity->getName()) ?>
                         </a>
                         <?php endif;?>
