+++ module/Activity/view/activity/activity/view.phtml
@@ -171,7 +171,7 @@ $this->headTitle($this->translate('Activities')); ?>
     <?php endif; ?>
 
     <?php if ($this->acl('activity_acl')->isAllowed('activity', 'viewDetails') && $activity->getCanSignUp()): ?>
-        <a href="<?= $this->url('organizer_activity/email', ['id' => $activity->getId()])?>"><?= $this->Translate('View email Adresses') ?></a> |
-        <a href="<?= $this->url('organizer_activity/export', ['id' => $activity->getId()])?>"><?= $this->translate('Export subscriptions') ?></a>
+        <a href="<?= $this->url('activity_admin/email', ['id' => $activity->getId()])?>"><?= $this->Translate('View email Adresses') ?></a> |
+        <a href="<?= $this->url('activity_admin/export', ['id' => $activity->getId()])?>"><?= $this->translate('Export subscriptions') ?></a>
     <?php endif; ?>
 </div>
