+++ module/Activity/view/activity/admin-approval/queue-approved.phtml
@@ -3,6 +3,6 @@
 echo $this->partial('activity/activity/list.phtml', [
     'activities' => $activities,
     'admin' => true,
-    'route' => 'admin_activity/queue_approved',
+    'route' => 'activity_admin_approval/queue_approved',
 ]);
