+++ module/Activity/view/activity/admin-approval/queue.phtml
@@ -12,7 +12,7 @@
 echo $this->partial('activity/activity/list.phtml', [
     'activities' => $unapprovedActivities,
     'admin' => true,
-    'moreLink' => $moreUnapprovedActivities ? $this->url('admin_activity/queue_unapproved', ['page' => 1]) : null,
+    'moreLink' => $moreUnapprovedActivities ? $this->url('activity_admin_approval/queue_unapproved', ['page' => 1]) : null,
 ]);
 ?>
 
@@ -21,7 +21,7 @@ echo $this->partial('activity/activity/list.phtml', [
 echo $this->partial('activity/activity/list.phtml', [
     'activities' => $approvedActivities,
     'admin' => true,
-    'moreLink' => $moreApprovedActivities ? $this->url('admin_activity/queue_approved', ['page' => 1]) : null,
+    'moreLink' => $moreApprovedActivities ? $this->url('activity_admin_approval/queue_approved', ['page' => 1]) : null,
 ]);
 ?>
 
@@ -30,6 +30,6 @@ echo $this->partial('activity/activity/list.phtml', [
 echo $this->partial('activity/activity/list.phtml', [
     'activities' => $disapprovedActivities,
     'admin' => true,
-    'moreLink' => $moreDisapprovedActivities ? $this->url('admin_activity/queue_disapproved', ['page' => 1]) : null,
+    'moreLink' => $moreDisapprovedActivities ? $this->url('activity_admin_approval/queue_disapproved', ['page' => 1]) : null,
 ]);
 ?>
