--- module/Activity/view/activity/organizer/view.phtml
@@ -1,84 +1,0 @@
-<div class="container">
-<h1><?= $this->translate('Upcoming activities') ?></h1>
-<table class="table table-striped">
-    <thead>
-    <tr>
-        <th>#</th>
-        <th><?= $this->translate('Dutch name') ?></th>
-        <th><?= $this->translate('English name') ?></th>
-        <th><?= $this->translate('Start date') ?></th>
-        <th><?= $this->translate('Subscriptions') ?></th>        
-        <th><?= $this->translate('Organ') ?></th>        
-        <th><?= $this->translate('Submitter') ?></th>
-        <th><?= $this->translate('Approval status') ?></th>
-        <th></th>
-    </tr>
-    </thead>
-    <tbody>
-    <?php foreach ($upcomingActivities as $activity): ?>
-        <tr>
-            <td><?= $activity->getId() ?></td>
-            <td><?= $activity->getName() ?></td>
-            <td><?= $activity->getNameEn() ?></td>
-            <td><?= $activity->getBeginTime()->format('Y-m-d') ?></td>
-            <td><?= is_null($activity->getSignups()) ? 0 : count($activity->getSignups()) ?></td>
-            <td><?= is_null($activity->getOrgan()) ? $this->translate('None') : $activity->getOrgan()->getAbbr()  ?></td>
-            <td><?= $activity->getCreator()->getMember()->getFullName() ?></td>
-            <td><?= $activity->getStatus() . ($activity->getUpdateProposal()->count() === 0 ? '' : $this->translate(' (Update pending)')) ?></td>
-            <td>
-                <a href="<?= $this->url('organizer_activity/update', ['id' => $activity->getId()]) ?>"><?= $this->translate('Update') ?></a>
-            </td>
-        </tr>
-    <?php endforeach; ?>
-    </tbody>
-</table>
-
-
-<h1><?= $this->translate('Old activities') ?></h1>
-<?= $this->paginationControl(
-    $oldActivityPaginator,
-    // the scrolling style
-    'elastic',
-    'paginator/default',
-    [
-        'path' => $this->url('organizer_activity')
-    ]
-)
-?>
-<?php if (!empty($oldActivityPaginator)): ?>
-<table class="table table-striped">
-    <thead>
-    <tr>
-        <th>#</th>
-        <th><?= $this->translate('Dutch name') ?></th>
-        <th><?= $this->translate('English name') ?></th>
-        <th><?= $this->translate('Start date') ?></th>
-        <th><?= $this->translate('Subscriptions') ?></th>        
-        <th><?= $this->translate('Organ') ?></th>        
-        <th><?= $this->translate('Submitter') ?></th>
-        <th><?= $this->translate('Approval status') ?></th>
-        <th></th>
-    </tr>
-    </thead>
-    <tbody>
-
-    <?php foreach($oldActivityPaginator as $activity): ?>
-        <tr>
-            <td><?= $activity->getId() ?></td>
-            <td><?= $activity->getName() ?></td>
-            <td><?= $activity->getNameEn() ?></td>
-            <td><?= $activity->getBeginTime()->format('Y-m-d') ?></td>
-            <td><?= is_null($activity->getSignups()) ? 0 : count($activity->getSignups()) ?></td>
-            <td><?= is_null($activity->getOrgan()) ? $this->translate('None') : $activity->getOrgan()->getAbbr()  ?></td>
-            <td><?= $activity->getCreator()->getMember()->getFullName() ?></td>
-            <td><?= $activity->getStatus() . ($activity->getUpdateProposal()->count() === 0 ? '' : $this->translate('(Update pending)')) ?></td>
-            <td>
-                <a href="<?= $this->url('organizer_activity/update', ['id' => $activity->getId()]) ?>"><?= $this->translate('Update') ?></a>
-            </td>
-        </tr>
-    <?php endforeach;?>
-    </tbody>
-</table>
-        
-<?php endif; ?>
-</div>
