+++ module/Activity/view/email/activity.phtml
@@ -34,7 +34,7 @@ Er is een nieuwe activiteit aangemaakt op de GEWIS website.
     </tr>
 </table>
 <br/>
-Klik <a href="<?= $this->serverUrl($this->url('admin_activity/view', ['id' => $activity->getId()])) ?>">hier</a> om deze activiteit goed/af te keuren.
+Klik <a href="<?= $this->serverUrl($this->url('activity_admin_approval/view', ['id' => $activity->getId()])) ?>">hier</a> om deze activiteit goed/af te keuren.
 <br/><br/>
 
 Met vriendelijke groeten,<br />
@@ -74,7 +74,7 @@ A new activity was created at the GEWIS website. See below for details.
     </tr>
 </table>
 <br/>
-Click <a href="<?= $this->serverUrl($this->url('admin_activity/view', ['id' => $activity->getId()])) ?>">here</a> to approve/disapprove this activity.
+Click <a href="<?= $this->serverUrl($this->url('activity_admin_approval/view', ['id' => $activity->getId()])) ?>">here</a> to approve/disapprove this activity.
 <br/><br/>
 
 Kind Regards, <br />
