+++ module/Application/view/partial/admin-nav.phtml
@@ -7,7 +7,7 @@
     <ul class="nav nav-pills head-menu">
         <?php if ($this->acl('activity_acl')->isAllowed('activity', 'approve')): ?>
         <li>
-            <a href="<?= $this->url('admin_activity') ?>">
+            <a href="<?= $this->url('activity_admin') ?>">
                 <?= $this->translate('Activities') ?>
             </a>
         </li>
